Added flare integration.
This allows the flame engine to be used in conjunction with flare and be
able to launch flare with the most recently published batch script to be
loaded at startup. In order to configure the flame engine to run flare,
the engine instance must be named 'tk-flare' rather than 'tk-flame'.