#27121 Backburner python executable now passed explicitly via bootstrap.
Previous, the python executable was identified via sys.executable, but
this returned the path to the binary on macosx rather than a python
interpreter. This change ensures that the associated python interpreter
is always used by backburner. All logic for calculating this sits in the
bootstrap, meaning that the engine is unaware of the organization of the
various binaries. This is analoguous to other more complex engines.