[minor] Added flame_major_version property and improved user name logic.
The default hook which works out which username to use when launching
flame now includes the major version as part of the name (e.g. "John
Smith (v2015)"). This is to ensure smooth transition across major
versions of Flame - user accounts are not compatible across major
versions so this is a way to ensure that there is a functioning user for
each DCC generation.