[minor] Regression fixes for Flame 2015.x
It turns out that some of the very minor changes done in the UI to
present an improved project creatio legacy UI is causing QA regressions
that are hard to pinpoint and test. In the interest of speed, this
change reverts those improvements. Flame integration for versions prior
to 2016.1 is now consistent with earlier versions of the flame engine
code, with the goal to eliminating subtle wiretap regressions.