#37898 Adds commands to flame menu by default
This is an improvement to the existing `context_commands` option in flame, where you can add custom commands to the menu (similar to the way for example Hiero works). This extends the flame engine to work more seamlessly with Toolkit and behave more like other engines; any registered command (except if it is of of type context_menu) is now automatically added to the flame menu. This makes flame more like maya and nuke where things will just automatically appear on the menu as soon as you have installed an app.

Closes #14.