Do not show "Load in new Batch Group" option is no batch file present
JIRA: SMOK-48548
"Load in new Batch Group" in SG loader should not be available on Movie export.

Refactor _add_batch_group_from_shot() to be able to call it partially from the
generate_actions() hook. This allow the action to be added only if the batch
setup is present.

While at it, I've exposed a bit more the error so when loading a batch
setup to a file that do not exists trigger an popup instead of just a error
in the log.

Fixed also some easy errors reported by PyLint around the code I've modified.