Return thubmnail generation job ids (#74)
JIRA: SMOK-49295
Use PyExporter in tk-flame-review

Return the Backburner jobs ids created by the Thumbnail generator so
dependent backburner job can be scheduled. This is required to integrate
the Thumbnail generator in tk-flame-review since we want to be able to
delete the created movie file when the thumbnail generation job is finished.