Support creating transcoding job of unwritten quicktime file
JIRA: SMOK-49295

Open Clip can open Quickime file with multiple handler.
When that happen, the file need to be there to know which one we want to use.

However, in the case of the transcoder, this Quicktime one is the one we
will want to use so remove the ambiguity.