Re-enable "Create Batch Group" otion in loader
JIRA: SMOK-49516
Do not create write file node to /var/tmp when creating batch from shotgun

Do not create write file node to /var/tmp when creating batch by default.

Setting the default to an empty string will cause an error which is what we
want.

Add Settings to configure the default value that can be used instead of an
env var.

Fix template parsing done for create_batch not working anymore

We need to take into accoutn the {flame.frame} token that was added after this
feature was disabled.

Fix the extension parsing not mapping correctly to the file type if a . was
present.

Also, added more supported types in the batch write file node.