Fix movie preview only have one frame (#76)
* Fix movie preview only have one frame

JIRA: SMOK-49295

We write an temporary Open Clip file pointing to the file that will be
exported by flame before so the preview job can be sent before the background
export job finishes.

However the file we were writting did not have a duration which implicitly
means that the nubmer of file is the number of frame (unless the media file
header can be read which is not the case when the export is still going).

This resulted in one frame movie file preview. The fix is to write the
duration in the Open Clip. I've added the fps at the same time to ensure proper
playback speed.

will be ignored, and an empty message aborts the commit. # # On branch
dev/braultj/movie-lenght # Your branch is up to date with 'origin/master'. # #
Changes to be committed: #	modified:   python/tk_flame/transcoder.py #