Go in batch before creating/loading batch group
JIRA: SMOK-49634

Some code assume you are in the batch tab before doing some action.

While this is bugs in Flame Python API, there is a legitime argment to be
able to control weither or not you go in batch when loading something in batch.

I've added a settings to control that behavior that is on by default.