Do not use multiple server if shared backburner tmp is /tmp (#83)
JIRA: SMOK-50909
backburner jobs submitted during shotgun publish get distributed to multiple backburner servers

When using a local path, we do not want the job created to be distributed
by BackBurner on all available servers, we want to restrict them to local paths.

When a server list or group is passed however, the backburner_shared_temp
settings must point to an actual shared path.

There is no easy way to detect what is shared or not. However using /tmp
by default causes problems.

This branch check and errors out if the default temporary folder is used
(which is /tmp most of the time) and a server list and/or a server group
has been defined since this will usually result in errors.

This branch also force local host as the server if the default temporary
folder is used and no server has been forces by the configuration since
this is the best way to ensure that the job will work.

Fix the job description that were a bit overzellous in the sanitization.

Corrected some issue flagged by pylint in modified file while at it.

python tempfile.gettempdir() can return something a bit unusual on mac.
Be sure to check the usual suspect too.