Support more asset types in publisher (#86)
JIRA: SMOK-51359
shot info is missing in Shotgun when published if "Create Open Clip" is disabled

The publisher was only considering the batchOpenClip as assert to create/update
the shots and cut items. It should be considering all possible assets.

Update the code to generate the thumbnail to correctly ignore asset type that
do not have video media.

Address simple whitespace issue found with pylint.