Improve backburner job name sanitization
JIRA: SMOK-51684
Shotgun can submit backburner jobs with a job name that is too long, error on submission

The maximum job name is 100 characters. However, we want to limit a bit lower
than than since duplicate job name will have a _### appended to the end of the
job name.

Create a method in the enging to do that sanitization to avoid code duplication
and support both a prefix and suffix to the job name that won't be ellipsed
unless we overflow too much.

Address some pylint warning/refactor issues.