Fix assetType not correctly defined when rendering in batch
JIRA: SMOK-51794
Shotgun batch render do not set assetType correctly

The assetType for everything produced by a batch render was flagged as a video
asset which is obviously wrong and cause the publisher to try to produce
thumbnail for the batch file.

Fix some line too long warning at the same time.