Improve support for machine having FQDN in host name
JIRA: FLME-58273

Use socket.gethostname() since it will return the hostname with .local on mac.
Keep the .local for the machien name since this is what backburner expect.
Strip the rest of the hostname.

Improve code that lookup for local path by only checking if the path starts with
temporary folder instead of contain.

Simiarly, check if the server starts with the localhost host name to support
server name with -# suffix.