Avoid refreshing credentials if not expired
JIRA: FLME-58463
ShotGrid: tk-flame reconnect to shotgun server each time a command line is sent

We refresh credentials before sending command line jobs since command line jobs
will be executed in a context where that will be impossible to show the UI
to validate the user crendentials.

This however can be quite long since it communicate with the ShotGrid server.
Only do it if credentials are expired.