Fix selection of proper application path for version extraction
JIRA: FLME-59475

Using realpath is problematic since paths in the /opt/Autodesk can daisy
chaing symlink and move to install structure that are not properly versionned.

Instead do the symlink resolving manually to dig as far as we actually need.

Also to do assume that the .app is a Flame since it can be a Flame, Flare or
Flame assist.