Improve multi-version support for user creation
JIRA: FLME-67427
Flame does not automagically load the selected project when launched from Flow Desktop

Use the wiretap command tools with the application version instead of using the Python API
since the python API will always use the latest version of the IFFFS wiretap server
when spawning a session.

This means, that in order to be able to create a user, we need to use a pre-2025 wiretap
server session which will be triggered by a pre-2025 wiretap command line tool.